tasks/property_tiles_info/requirements.txt
functions_framework==3.8.2
google-cloud-bigquery==3.31.0
google-cloud-storage==3.1.0
